+++ CHANGELOG.md
@@ -13,6 +13,24 @@ Nomenclature :
 
 ### Unreleased
 
+#### Added
+
+### [0.1.1] - 2023-03-29
+
+#### Added
+
+- Added GET /fixtures-inventory
+- Added fixture-inventory table with some data (014)
+- Added GET /fixtures-inventory/:id
+- Added POST /fixtures-inventory
+- Added PUT /fixtures-inventory/:id
+- Added DELETE /fixtures-inventory/:id
+- Added unit test for method getFixtureInventoryById() of the service
+
+#### Updated
+
+- Updated unit property test to integrate fixture inventory id
+
 ### [0.1.0] - 2023-03-27
 
 #### Added